move to validate.go
