+++ flow/connectors/postgres/validate.go
@@ -8,55 +8,55 @@ import (
 	"strings"
 
 	"github.com/jackc/pgx/v5"
+
+	"github.com/PeerDB-io/peer-flow/connectors/utils"
+	"github.com/PeerDB-io/peer-flow/shared"
 )
 
-func (c *PostgresConnector) CheckSourceTables(ctx context.Context, tableNames []string, pubName string) error {
+func (c *PostgresConnector) CheckSourceTables(ctx context.Context,
+	tableNames []*utils.SchemaTable, pubName string,
+) error {
 	if c.conn == nil {
 		return errors.New("check tables: conn is nil")
 	}
 
 	// Check that we can select from all tables
 	tableArr := make([]string, 0, len(tableNames))
-	for _, table := range tableNames {
+	for _, parsedTable := range tableNames {
 		var row pgx.Row
-		schemaName, tableName, found := strings.Cut(table, ".")
-		if !found {
-			return fmt.Errorf("invalid source table identifier: %s", table)
-		}
-
-		tableArr = append(tableArr, fmt.Sprintf(`(%s::text, %s::text)`, QuoteLiteral(schemaName), QuoteLiteral(tableName)))
+		tableArr = append(tableArr, fmt.Sprintf(`(%s::text, %s::text)`,
+			QuoteLiteral(parsedTable.Schema), QuoteLiteral(parsedTable.Table)))
 		err := c.conn.QueryRow(ctx,
-			fmt.Sprintf("SELECT * FROM %s.%s LIMIT 0;", QuoteIdentifier(schemaName), QuoteIdentifier(tableName))).Scan(&row)
+			fmt.Sprintf("SELECT * FROM %s.%s LIMIT 0;",
+				QuoteIdentifier(parsedTable.Schema), QuoteIdentifier(parsedTable.Table))).Scan(&row)
 		if err != nil && err != pgx.ErrNoRows {
 			return err
 		}
 	}
 
 	tableStr := strings.Join(tableArr, ",")
-	if pubName != "" {
-		// Check if publication exists
-		err := c.conn.QueryRow(ctx, "SELECT pubname FROM pg_publication WHERE pubname=$1", pubName).Scan(nil)
-		if err != nil {
-			if err == pgx.ErrNoRows {
-				return fmt.Errorf("publication does not exist: %s", pubName)
-			}
-			return fmt.Errorf("error while checking for publication existence: %w", err)
+	// Check if publication exists
+	err := c.conn.QueryRow(ctx, "SELECT pubname FROM pg_publication WHERE pubname=$1", pubName).Scan(nil)
+	if err != nil {
+		if err == pgx.ErrNoRows {
+			return fmt.Errorf("publication does not exist: %s", pubName)
 		}
+		return fmt.Errorf("error while checking for publication existence: %w", err)
+	}
 
-		// Check if tables belong to publication
-		var pubTableCount int
-		err = c.conn.QueryRow(ctx, fmt.Sprintf(`
+	// Check if tables belong to publication
+	var pubTableCount int
+	err = c.conn.QueryRow(ctx, fmt.Sprintf(`
 		with source_table_components (sname, tname) as (values %s)
 		select COUNT(DISTINCT(schemaname,tablename)) from pg_publication_tables
 		INNER JOIN source_table_components stc
 		ON schemaname=stc.sname and tablename=stc.tname where pubname=$1;`, tableStr), pubName).Scan(&pubTableCount)
-		if err != nil {
-			return err
-		}
+	if err != nil {
+		return err
+	}
 
-		if pubTableCount != len(tableNames) {
-			return errors.New("not all tables belong to publication")
-		}
+	if pubTableCount != len(tableNames) {
+		return errors.New("not all tables belong to publication")
 	}
 
 	return nil
@@ -109,6 +109,52 @@ func (c *PostgresConnector) CheckReplicationPermissions(ctx context.Context, use
 		return errors.New("max_wal_senders must be at least 2")
 	}
 
+	return nil
+}
+
+func (c *PostgresConnector) CheckPublicationPermission(ctx context.Context, tableNameString string) error {
+	publication := "_PEERDB_DUMMY_PUBLICATION_" + shared.RandomString(4)
+	// check and enable publish_via_partition_root
+	supportsPubViaRoot, _, err := c.MajorVersionCheck(ctx, POSTGRES_13)
+	if err != nil {
+		return fmt.Errorf("error checking Postgres version: %w", err)
+	}
+	var pubViaRootString string
+	if supportsPubViaRoot {
+		pubViaRootString = "WITH(publish_via_partition_root=true)"
+	}
+	tx, err := c.conn.Begin(ctx)
+	if err != nil {
+		return fmt.Errorf("error starting transaction: %w", err)
+	}
+	defer func() {
+		err := tx.Rollback(ctx)
+		if err != nil && err != pgx.ErrTxClosed {
+			c.logger.Error("[validate publication create] failed to rollback transaction", "error", err)
+		}
+	}()
+
+	// Create the publication
+	createStmt := fmt.Sprintf("CREATE PUBLICATION %s FOR TABLE %s %s",
+		publication, tableNameString, pubViaRootString)
+	_, err = tx.Exec(ctx, createStmt)
+	if err != nil {
+		return fmt.Errorf("it will not be possible to create a publication for selected tables: %w", err)
+	}
+
+	// Drop the publication
+	dropStmt := "DROP PUBLICATION IF EXISTS " + publication
+	_, err = tx.Exec(ctx, dropStmt)
+	if err != nil {
+		return fmt.Errorf("it will not be possible to drop the publication created for this mirror: %w",
+			err)
+	}
+
+	// commit transaction
+	err = tx.Commit(ctx)
+	if err != nil {
+		return fmt.Errorf("unable to validate publication create permission: %w", err)
+	}
 	return nil
 }
